added structure to the launcher
