+++ crates/modules/launcher/src/application.rs
@@ -1,0 +1,1 @@
+pub mod app_search {}
