+++ crates/modules/launcher/src/equation.rs
@@ -1,0 +1,1 @@
+pub mod equate {}
