+++ crates/modules/launcher/src/search.rs
@@ -1,0 +1,10 @@
+pub mod search {
+    pub enum Pattern {
+        Application,
+        Equation,
+    }
+
+    pub fn pattern_match(match_string: &String) -> Pattern {
+        return Pattern::Application;
+    }
+}
