+++ crates/modules/launcher/src/wsearch.rs
@@ -1,0 +1,1 @@
+pub mod web_search {}
